hid-decode: read libinput record yaml files
`libinput record` now also integrate the hid report descriptors of the
device when the device is a hid one.

Instead of relying on custom scripts to convert the files into
`hid-recorder` compatible ones, teach `hid-decode` to parse the provided
report descriptors directly.

Note, this adds a new dependency on PyYaml, but it should be common enough
on all distributions now.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>