uhid: pull all events off our fds during dispatch()
Previously we'd only ever process one event from the fds. Or rather - it was
implementation-defined based on the functions that were called. None of those
were pulling data off until it was gone, so let's wrap things into a loop here
and keep calling until we're properly out of data.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>