tests: base: do not care about the SYN_REPORT value
in most cases, the SYN_REPORT value is 0, but it can happen that we get
a value of 1. This happens with the ITE keyboards, when simulatenously
pressing 6 keys at a time. We are basically filling the output buffer
of the input stack, so the kernel adds a SYN_REPORT with 1, except that
the next event is SYN_REPORT 0, which gets dropped.

For everyone's sake, drop the value handling, it will be better.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>