Refactor replay to use known devices
If there is a specific class already created for the device being
replayed then assign it from that list. This helps for devices which
do not use /dev/input nodes, which hangs the program. The device with
which this has been tested so far is the dual shock 3 controller. This
solves the problem of the /dev/hidraw device not being destroyed when
rpelay hangs.