BaseDevice: Make 'match_evdev_rule' more lenient
At the moment the function offers no middle-ground between
returning "true" and associating an evdev device with a specific
type, or returning "false" and associating it with *no* types.

The Wacom tests require us to have access to both e.g. Pen and
Pad devices, but we need to be careful to ensure that the pen
evdev device is actually associated with the "Pen" type, and
pad device with the "Pad" type. The "udev_input_event" function
does a pretty good job with this normally, but because all pad
devices are also pen devices, things go a little off the rails.

This commit allows the caller to return 'false' for only the
specific combinations of device and application that don't
make sense, but still return 'true' in other situations.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>