tests: wacom: Implement test for partial/empty pad syncs
The PTH-460 has its HID descriptor implemented slightly differently
to its siblings the 660 and 860. Specifically, each of the functions
of the pad are encapsulated in their own physical collections. This
causes a little bit of trouble for the current Wacom driver which
typically treats multiple collections of the same tool as multiple
snapshots of a single tool's state (mainly to support BT devices
that send e.g. 5 pen updates in a single packet).

We check to make sure that we don't get partial/empty pad syncs
when they shouldn't exist. The easiest way to trigger this behavior
is to simultaneously use the touchring and expresskeys since they
are in different collections.

This behavior was fixed in the upstream Linux kernel by commit
d4b8efeb46d9 ("HID: wacom: generic: Correct pad syncing").

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>