sony-gamepad: add initial DualShock 4 Bluetooth implementation
The DualShock 4 in Bluetooth mode is a complicated device. In its default
configuration it reports data through "reportID 1", which is a small input
report which only exposes axis and button state. None of the advanced features
such as touchpad, motion sensors are reported to this report.

When feature report '0x05' is used during startup of hid-sony, this retrieves
calibration data, but this also switches the device to a different power mode
and switches default input report to reportID 17. This report uses a vendor
specific layout, but is a 'super set of reportID 1' which also adds touchpad,
motion sensors, battery info as well as a CRC.

A custom create_report implementation is used to create input data using
reportID 17 and also calculate the CRC.

Signed-off-by: Roderick Colenbrander <roderick.colenbrander@sony.com>