sony-gamepad: fix incorrect feature report data for PS4ControllerUSB
The motion sensor calibration data in feature report 0x02 is incorrect.
The bytes for the gyro_pitch_plus and gyro_pitch_minus data are swapped.
It caused gyro pitch (y-axis) data to have the wrong sign.
This was likely due to a manually created feature report by someone,
but went unnoticed until now.

Signed-off-by: Roderick Colenbrander <roderick.colenbrander@sony.com>