sony-gamepad: Add touchpad support for PS5 controllers
Add support for PS5 touchpads. The touchpad is very similar to PS4
touchpad except there is no history this time and the report layout
is slightly different e.g. timestamp is at the end instead of the
start.

The touchpad support for PS5Controller is implemented in a similar
way to PS4Controller. Main difference is that there is a new
PS5TouchReport class to handle the differences in report layout.

Signed-off-by: Roderick Colenbrander <roderick.colenbrander@sony.com>