Use sphinxcontrib-apidoc to build the API
Instead of relying on an external makefile to build this, let's use the
python-specific approach. setuptools already has this integrated:

   python setup.py build_sphinx

We only need to add the apidoc extension to conf.py and done.

Paths are relative to conf.py, so our output path of '.' dumps the generated
.rst files into the doc/source directory. Ideally we'd want the generated
files inside the 'build' directory but I have yet to figure out how to get
the 'build' directory from within the script.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>