hid: handle a vendor-specific collection type
A collection type of 0x80-0xff is VENDOR_DEFINED which we handled in one
location but not the other. This causes an exception when printing the human
description of a report.

Fix this by moving the collection type handling into a helper function, to
be used the same way in both places we require it.

Fixes #25

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>