hid: fix byte count calculation when parsing human descriptions
It's unclear why this was handled this way but it produced wrong
calculations for items at the bit boundary - all those got bumped up to the
next size (e.g. vendor usage 0xffff ended up being 0x0000ffff in the report
because it got bumped to the next size).

The number of bytes needed is simply whatever the actual value requires, no
need to count the length of the hex string.

This produces the wrong value for cases where extra bytes are present in the
original file descriptor, e.g. this case:
  # 0x26, 0xff, 0x00,              //   Logical Maximum (255)
This line reconverted would output a single-byte only
  # 0x25, 0xff                     //   Logical Maximum (255)

Without parsing the byte count at the end of the outputted line, this isn't
something that can be fixed here.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>