tests: switch the udev rule creator to a class
We can make use of python's 'with' context manager to create an object that
automatically removes itself when we're done with it. This removes the need
to have a setup/teardown API imported from the other file.

The context manager is preferable over overriding __del__ as we cannot
predict when the latter is actually called.

This also switches from using uuids for tmpfile generation to the specific
function designed for this.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>