tests: modify the udev rules setup to only start on the second context
We don't need udev rules until we run a test that creates a test device.
So we modify our class to set up the udev rule on the second context
entered (first is session, second is the current test case) and clean up
once the last context (== session context) is exited.

This way we never mess with udev rules if we only run tests without test
devices.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>