gitlab CI: hook up a qemu-based pytest run
Allowed to fail because we're not tracking the upstream kernel so some tests
are expected to fail. But this will at least detect severe breakages that
pass the existing flake/setup.py checks.

Before we can run the test though we need to modprobe all of the hid modules.
The ITE keyboard has its own kernel module and that causes device creation
to fail, specifically: "It turns out that this device loads the hid_ite
kernel module. So what happens is that we create the hid device, hid-generic
sees it, presents the nodes to the userspace, but then udev loads hid-ite,
which kills the input nodes attached to hid-generic and everything related
to it. The next attempt at loading the device works fine, because hid-ite is
there already."
See https://gitlab.freedesktop.org/libevdev/hid-tools/-/issues/29#note_766562
Simply modprobing all modules before we run the test fixes this issue. This
is a separate script to avoid the escaping of subshells otherwise needed to
get to the kernel module path.

Fixes #29

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>