test: add a test to print and re-read a report descriptor
Printing the human form from an rdesc and re-reading that form should yield
the same report descriptor as before. For the test we use the same rdesc as
for the parsing test above, with some Logical Maximums fixed manually to
only use the number of bytes they should use.

Since we don't (yet) parse the byte number at the end of the human output,
we cannot guess that 0xff wants 2 bytes in the item.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>