Sony: PS4: wait for the power_supply to be added too
We have a test that checks for the battery, but we present the device
whenever the inputs and leds are ready. This does not impact normally
because the battery is presented by the kernel before the power_supply
is added, but if we fail at adding the power_supply, we are failing
with an incorrect message.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>