hid: fix check of the report size
As the comment suggests, if a report (i.e. the event data) is larger than it
should be according to the ReportDescriptor, that's fine, we'll just end up
ignoring the bits.

This fixes issues with the stylus recordings seen in
https://gitlab.freedesktop.org/libinput/libinput/-/merge_requests/558#note_792834
That device sends 32 bytes for report ID 8 (13 bytes in the descriptor)

Fixes #32

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>