tests: tablet: add initial support for testing pen devices
This is mostly incomplete, but should allow to move forward on the LKML.

The thing is I don't have any of these devices, and I don't know If I am
doing the right thing here.

There are also a bunch of commented out devices. These are devices that
are exposing a Stylus from one of their collections, but do not work
straight out of the box:
- egalax devices are exporting a 'Touch Screen' application
- Irtouch_6615_0070 does not set the BTN_TOOL_PEN (though it has in-range)
- atmel_03eb_8409 has an 'undefined' application for its pen

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>