tests: add a new set of tests to attempt to crash the kernel
The CI will likely fail on this, we need an updated kernel.

This was raised on the LKML and privately: it is very easy to crash
the kernel with uhid by picking up one of the USB modalias reserved
for those drivers.

Retrieve all of those USB VID/PIDs, and attempt at binding them.
If there is no crash, we are good.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>