hid-recorder: primary use the logical name for arrays
Seen on the USI digitizers: the application is `Digitizer`, but the
logical is `Preferred Line Style`. This logical is actually better
so use that as the debug output.

Furthermore, this makes it way easier to create keyboard arrays
as we can now use `data.keyboard` instead of
`setattr(data, '0x70000')`, which is cumbersome to say the least.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>