tests: tablets: implement better and more covering testing
The tests were manually written and it was quite cumbersome to add
a new state.

However, the state machine we expect from a Pen device is given in the
spec from Microsoft at
https://docs.microsoft.com/en-us/windows-hardware/design/component-guidelines/windows-pen-states

Use this as a base to:
- express each of these states as an enum value
- have a generic hook to move from one state to one other
- have a generic facility to test the transition between tests

Suddenly, we can remove the manual tests for hovering/touch, and emulate
all sort of transitions, the valid ones from the MS spec, but also
some others, in case we have a device that is not playing correctly.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>