tests: tablets: look for valid transitions at the evdev level
For instance, the HID state machine authorizes to move from in-range
to in-range-with-the-intent-to-erase. However, this transition confuses
the wacom xf86 driver.

Add a set of valid transitions from an evdev point of view, decompose
the received events in a series of EV_SYN chunks, and validate that each
chunk moves to a new valid state.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>