base_device: do not rely on udev for fetching input objects
Instead of parsing all devices one by one, we can rely on udev to only
know when the device has been initialized in the kernel through the "bind"
property.
Now we know that no more input nodes will be added.

Note: this rely on the fact that the module driving the uhid device has
been pre-loaded before the tests or that it was builtin the vmlinux

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>