tests_kernel: keep the application matches in the tests_kernel subdir
Given that test_kernel are going to be shipped in the kernel tree, we need
a way from there to edit the application matches when we add a new test.

This way, we don't have to sync hid-tools and the kernel tree when we
introduce a new test where the application match is not working with the
current implementation.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>