tests kernel: ensure we pre-load the tested module before testing
The recent changes in how the device is considered as ready implies that
udev doesn't have to load an external module for the device to be ready.

Manually tag which kernel module needs to be present in each test, and
if it is not, skip the test.

This also allows for not waiting 5 seconds for a device to be ready when
the kernel module is not compiled in.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>