dist: convert to hatch
Instead of trying to fix the now not-entirely deprecated setup.py, convert
the project to `pyproject.toml`, and use `hatch` as the build tool.

`hatch` allows a couple of tricks like automatically bumping the version with
`hatch version minor`, or running the tests, even if I haven't pushed it that
far.

Generating the man pages required a custom hook in wheel and force the
inclusion of the markdown sources in the source dist.

Fixes: #45

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>